refactor: 페이지 컴포넌트 디렉토리로 이동
